+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.2.0
+- Enhanced Matches Tree View
+  - Multiple match groups can now be set
+  - Match group shows document name and match text
+  - Preference can be set for limit of match groups to show
+- Recent Edits Section
+  - Shows recent edit locations
+  - Preference can be set for what is considered the span of a single edit
+  - Preference can be set for how many edits to show
 
 ## 1.1.0
 - Added a tree view to show current find matches